replaced some NGL_OUT by NGL_LOG fixed windows time converters
git-svn-id: svn://libnui.net/nui/trunk/nui3@4942 9db30882-7939-4ce0-a36a-45f8a15c01b5

+++ include/nglTime.h
@@ -162,7 +162,7 @@ private:
   double mValue;
 #ifdef _WIN32_
   static __int64 mLaunchValue;
-  static int32 mLaunchDate;
+  static int64 mLaunchDate;
   static __int64   mTimerFrequency;
   static double GetTime();
   friend class nglTimer;
